Running scripts from terminal now works.
However, you still cannot eg. remove old script through terminal.
Make terminal program for that.

+++ src/driver/motor/inc/motor.h
@@ -46,7 +46,7 @@ typedef struct motor_t
 	uint32_t delay;
 	
 	/* Current position in step value list. */
-	uint8_t step;
+	int32_t step;
 	
 	/* This counts how many ms have gone since last step. */
 	uint32_t delay_counter;
